Resolved 429 status code issue
Due to implementation of process pool
api got too many requets due to which
it returned 429 status code with no
content in response.

Also there wasn't any retry-after header
in the response so I made the worker sleep
for 1 second before retrying. It makes a
maximum of 3 tries before failing.